feat: ノート移行ダイアログにアセット削除オプションを追加
MigrateAssetToNoteにdeleteAssetパラメータを追加し、移行時にアセットを
残すか削除するかを選択可能にした。ダイアログのメッセージを簡潔にし、
メタ画像設定ダイアログと同様のチェックボックスUIに統一。

Glory to mankind.

+++ asset.go
@@ -398,8 +398,8 @@ func (b *Binder) GetUnpublishedAssets() ([]*json.Asset, error) {
 }
 
 // MigrateAssetToNote はテキストアセットをノートに移行する。
-// アセットの内容をノートのMarkdownとして登録し、元のアセットを削除する。
-func (b *Binder) MigrateAssetToNote(assetId string) (*json.Note, error) {
+// アセットの内容をノートのMarkdownとして登録し、deleteAsset が true の場合は元のアセットを削除する。
+func (b *Binder) MigrateAssetToNote(assetId string, deleteAsset bool) (*json.Note, error) {
 
 	if b == nil {
 		return nil, EmptyError
@@ -468,26 +468,30 @@ func (b *Binder) MigrateAssetToNote(assetId string) (*json.Note, error) {
 		return nil, xerrors.Errorf("createStructure() error: %w", err)
 	}
 
-	// アセットをDBから削除
-	err = b.db.DeleteAsset(assetId)
-	if err != nil {
-		return nil, xerrors.Errorf("db.DeleteAsset() error: %w", err)
-	}
-	err = b.db.DeleteStructure(assetId)
-	if err != nil {
-		return nil, xerrors.Errorf("db.DeleteStructure() error: %w", err)
-	}
-
-	// アセットファイルを削除
-	assetFiles, err := b.fileSystem.DeleteAsset(a)
-	if err != nil {
-		return nil, xerrors.Errorf("fs.DeleteAsset() error: %w", err)
-	}
-
 	// 全変更をコミット
 	files := []string{fn}
-	files = append(files, assetFiles...)
-	files = append(files, fs.NoteTableFile(), fs.AssetTableFile(), fs.StructureTableFile())
+	files = append(files, fs.NoteTableFile(), fs.StructureTableFile())
+
+	if deleteAsset {
+		// アセットをDBから削除
+		err = b.db.DeleteAsset(assetId)
+		if err != nil {
+			return nil, xerrors.Errorf("db.DeleteAsset() error: %w", err)
+		}
+		err = b.db.DeleteStructure(assetId)
+		if err != nil {
+			return nil, xerrors.Errorf("db.DeleteStructure() error: %w", err)
+		}
+
+		// アセットファイルを削除
+		assetFiles, err := b.fileSystem.DeleteAsset(a)
+		if err != nil {
+			return nil, xerrors.Errorf("fs.DeleteAsset() error: %w", err)
+		}
+		files = append(files, assetFiles...)
+		files = append(files, fs.AssetTableFile())
+	}
+
 	err = b.fileSystem.Commit(fs.M("Migrate Asset to Note", n.Name), files...)
 	if err != nil {
 		return nil, xerrors.Errorf("Commit() error: %w", err)
